+++ src/deployers/file.js
@@ -3,6 +3,7 @@ const gulp = require('gulp');
 const gutil = require('gulp-util');
 const using = require('gulp-using');
 const changed = require('gulp-changed');
+const excludeGitignore = require('gulp-exclude-gitignore');
 
 class File extends Deployer {
 
@@ -12,12 +13,16 @@ class File extends Deployer {
 	 * @param {string} src
 	 * @param {string} dest
 	 * @param {Object} connection
+	 * @param {boolean} gitignore
 	 * @return {Object} stream
 	 */
-	deploy(src, dest, connection){
+	deploy(src, dest, connection, gitignore){
 
 		// get source file(s)
 		let s = gulp.src(src);
+
+		// exclude files ignored in .gitignore
+		if(gitignore) s = s.pipe(excludeGitignore()).on('error', gutil.log);
 
 		// deploy changed files only
 		s = s.pipe(changed(dest)).on('error', gutil.log);