+++ src/deployers/ftp.js
@@ -2,6 +2,7 @@ const Deployer = require('./deployer.js');
 const gulp = require('gulp');
 const gutil = require('gulp-util');
 const ftp = require('vinyl-ftp');
+const excludeGitignore = require('gulp-exclude-gitignore');
 
 class Ftp extends Deployer {
 
@@ -12,9 +13,10 @@ class Ftp extends Deployer {
 	 * @param {string} src
 	 * @param {string} dest
 	 * @param {Object} connection
+	 * @param {boolean} gitignore
 	 * @return {Object} stream
 	 */
-	deploy(src, dest, connection){
+	deploy(src, dest, connection, gitignore){
 
 		// no connection
 		if(!connection) throw new gutil.PluginError('Propeller FTP Deployer', `No FTP connection information given`);
@@ -27,6 +29,9 @@ class Ftp extends Deployer {
 
 		// get source file(s)
 		let s = gulp.src(src, { base: '.', buffer: false }).on('error', gutil.log);
+
+		// exclude files ignored in .gitignore
+		if(gitignore) s = s.pipe(excludeGitignore()).on('error', gutil.log);
 
 		// only upload newer files
 		s = s.pipe(connect.newer(dest)).on('error', gutil.log);